requirements.txt
pyyaml
py2exe
pyinstaller
pillow
pyscreenshot
packaging
pynput
pycryptodome
pyreadline
colorama
